requirements.txt
Flask==2.0.3
PyYAML==6.0

